[CS214] - Asst1: Added workloads ABCD with timing
Added new function implementations for workloads A, B, C, and D.

I've commented out the line that includes the "mymalloc.h" because I
wanted to do the implementation of the workloads independent of our
malloc implementation.

Our current implementation in mymalloc doesn't exactly work with the
current workloads but  memgrind *does* work with stdlib's malloc.

memgrind with

```
make memgrind
./memgrind.out
```