[CS214] - Asst1: Full A-D Workloads
Fixed a couple issues leading to segfaults and incorrect frees on memory

1. myfree was accidentally doing a comparison on the NULL character as
opposed to an assignment when dealing with a single byte free
2. Also in myfree we never assigned the beginning index of the block to
null resulting in odd behavior where we wouldn't free the beginnings of
blocks. leaving 'b's everywhere.
3. Fixed an issue in allocate_mem where I was accidentally accessing an
index which was 1 too high (just needed to subtract 1 from the index).
4. Also added amethod to print all of MEM_SIZE's metadata. Not used
anywhere yet though. Just in case

Also the makefile now has an option for 'make debug' which generates a
debuggable executable for memgrind.

-Z