[CS214] - Asst1: Errors on mallocing MEM_SIZE
Basically the algorithm for searching in isAllocated was incorrect. It
wouldn't allow us to search for a new starting point if we go over the
"MEM_SIZE" value and we don't actually get to malloc in a place we
should have.

In order to alleviate this problem I did a couple things:

1. Added a static int isallocated_search_size which is a variable that
is set within the isAllocated method. This variable tells us how far we
searched in isAllocated before halting the search due to bounds or
allocated memory blocks.

2. I added a new int called broke to isAllocated. This will tell us
after the loops whether or not we broke while searching for free memory.
This tells us how we should return and what to set
isallocated_search_start to.

3. I also modified the bounds of the isAllocated to stop once we get to
i - start >= size I believe we were doing an extra loop iteration
before.

4. I also renamed the count variable to malloc_count and added a new
method get_bytes_malloc() which returns an int for how many bytes were
malloc'd

5. I deleted the unused free_mem method.