[CS214] - First iteration of multi-process
Pretty basic fork-exec-wait in order to get multiple processes to change
the code.

Also I reread the direction and it says each worker much independently
access the file - so we'll need to change how we handle that in the
threaded version as well. I believe I'm implemented it correctly here.