[CS214] - Fixed some bugs & added some edge cases.
There was a problem with the main LOLS algorithm regarding it not
mallocing space if the letters only spanned 0-2 letters. I changed it
to malloc some space and copy the string to that malloc'd space so we
wouldn't get any double free's.

Added some checks to make sure there was content in the strings being
returned from read_file. If there wasn't we'd terminate the
process/thread. Along with those, I added a check to make sure that the
number of times attempted to split didn't exceed the amount of
characters in the string.

Created more test cases inside test_rle (without assertions) to make sure everything was working.

Fixed the bug in itoa(). Once the integer would exceed single digits,
it would need to be reversed.