[CS214] - No more leaks except for CompressR
I fixed an issue in append_string that valgrind brought up where we were
copying a string in one buffer into the the same exact buffer. Valgrind
wasn't happy about it so I fixed it.

I also added one to the size of the malloc'd compressed string in order
to stop us from doing an invalid read/write on a single byte.

Lastly, it turns out we don't actually need to call pthread_exit because
it resulted in resources not being gathered properly. Since we already
join I could simmply just remove the pthread_exit() statements.