maze generation (also viewable in game loop)
3 Methods

Random tile selection - based on a probability a tile is modified.

(Modified) Prims - builds a maze based on whether or not neighbors are walls or paths

DFS - Generates a maze by making steps of two in random directions (best method)

in the game loop DFS is currently the visible method.